+++ src_blockchain_p2p/bc_p2p_cmd_node/SyncHeaderHeightData.cpp
@@ -106,7 +106,7 @@ UnicodeString* SyncHeaderHeightData::toString() const noexcept {
 		}
 
 		ret->append(L"Height : ");
-		ret->append((int)height);
+		ret->append((int64_t)height);
 		ret->append(L"[");
 		{
 			UnicodeString* str = id->toString(); __STP(str);
